Update to remove redudant components.
